[completed] 1-trusting-trust part 3
Successfully inserted the login exploit into `login.c`. The key was
including both the compiler and login exploits in `attack.c`. Then,
generate the decimal representation of `attack.c` and put it in
`attack.h`. Compile `trojan-cc2` from the concatenation of `attack.h`
and `trojan-cc2.c`. At this point, `attack.c`, `attack.h`, and
`trojan-cc2.c` can all be thrown away.

Next, compile the `identity-cc.c` using `trojan-cc2`, which inserts both
exploits into the resulting `cc-attacked` binary. We then use the
`cc-attacked` binary to compile `identity-cc.c` into `cc-attacked2`,
which contains the login exploit.

In my understanding, we use `trojan-cc2` to inject the exploits once
into the compiler binary `cc-attacked`. At this point, all previous
steps can be thrown away and `cc-attacked` can (and will) be used as the
compiler that compiles the compiler source code. Because the exploits in
`cc-attacked` cannot be detected, they're propagated to `cc-attacked2`,
the recently built version of the compiler that has clean source code.
So even though the compiler's source code is clean, the login exploit is
still there.