[completed] 2-dynamic-code-gen part2
Added dynamic code generation to speed up the calling of interrupts
handlers. Basically a copy of `1-hello`. The one gotcha was that the
code array had to be static, otherwise the size of the jump from the
text segment to the stack was too large and overflowed the jump capacity
of the branch instruction. I need to add an assert in
`unix-side/armv6-insts.h` to protect against the issue, but that's for a
later commit.